Add initial draft of existing practice section
Adds the table presented at C++Now with the following modifications:
- adds Armadillo, another popular linear algebra library
- adds xtensor, a project which is related to xsimd but for arbitrarily sized
  multidimensional arrays
- removes the Python library examples

Adds a brief section on the difference between a SIMD wrapper and a DSL, as I
see it.

Adds a diff to compare an example presented in P1928R3 with the new code.

This draft does not introduce a name for the new element-wise comparison
functions. This should be addressed before submission.

I believe I need to elaborate more on why we are going against the overwhelming
majority of existing practice. The short reason is something like: "these
libraries are either DSLs, or implemented for people who are used to using a
DSL and not for seamless interoperability with the rest of the C++ standard
library or language".